Move tests for GET /ctms, matches for alternate ID
